requirements.txt
faster-whisper
openai
opencv-python
ffmpeg-python
python-dotenv
imageio-ffmpeg
torch>=2.6.0
transformers==4.55.0
sentencepiece
minio
yt-dlp
